Okay so its messy right now but you can pick up, look at, drop, lock, unlock items. Theoretically you could also open and close them but I dont know if you could actually.
You also need the correct key to unlock and lock doors now.

All items are saved to the player object, which is an instance of a class, which means if you end the program with the player holding something, it no longer is holding it (so beware of that during testing)

You cannot see anything in the room if nothing in that room emits light (the room itself, the items the player is holding, and the items in the room

I still need to implement the 'in' preposition

I changed all the structure of how the parser inputs data. The third list holds what the subject is (room_item, player_item, exit)

After every action, the room gets saved.